requirements.txt
Django==2.1
djangorestframework==3.9.2
PyMySQL==0.10.1
pytz==2020.4
